Refactored the scripting API files into a Python package. Each class is moved into a separate module. __init__.py holds the public API and does the GUI/platform abstraction.
